refactor: routes enum
